rebased client changes
