--- client/src/components/ChatGXY/chatTypes.ts
@@ -1,24 +1,0 @@
-import type { ActionSuggestion, AgentResponse } from "@/composables/agentActions";
-
-export interface ChatMessage {
-    id: string;
-    role: "user" | "assistant";
-    content: string;
-    timestamp: Date;
-    agentType?: string;
-    confidence?: string;
-    feedback?: "up" | "down" | null;
-    agentResponse?: AgentResponse;
-    suggestions?: ActionSuggestion[];
-    isSystemMessage?: boolean;
-}
-
-export interface ChatHistoryItem {
-    id: string;
-    query: string;
-    response: string;
-    agent_type: string;
-    agent_response?: AgentResponse;
-    timestamp: string;
-    feedback?: number | null;
-}
